+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/HandlerAdapters.java
@@ -1,7 +1,5 @@
 
-package com.techcourse;
-
-import webmvc.org.springframework.web.servlet.mvc.HandlerAdapter;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import java.util.ArrayList;
 import java.util.List;
